Modularize and reorganize code
Broke out Jackson ObjectMapper into a utility class. Changed
OneCallResponseMapper to be WeatherForecastMapper, accepting a
GeocodeResponse as well. Removed GeocodeResponseMapper and Geolocation
model for weather.

+++ src/main/java/com/weather/discordweather/client/mapquest/model/Coordinate.java
@@ -1,6 +1,6 @@
 package com.weather.discordweather.client.mapquest.model;
 
 public record Coordinate(
-    float lat,
-    float lng
+    double lat,
+    double lng
 ) {}
